+++ extensions/hide-files/src/hooks/hooks.ts
@@ -44,7 +44,7 @@ export const alertDialog = async (
   message: string,
   confirmTitle: string,
   confirmAction: () => void,
-  cancelAction: () => void
+  cancelAction?: () => void
 ) => {
   const options: Alert.Options = {
     title: title,