+++ extensions/hide-files/src/toggle-system-files-visibility.tsx
@@ -5,7 +5,7 @@ import { LocalStorageKey } from "./utils/constants";
 export default async () => {
   const localStorage = await LocalStorage.getItem<boolean>(LocalStorageKey.LOCAL_HIDE_TOGGLE);
   const currentState = typeof localStorage === "undefined" ? true : localStorage;
-  await showHUD(`System hidden files are ${currentState ? "hidden" : "shown"}`);
+  await showHUD(`${currentState ? "Hidden" : "Unhidden"} system hidden files`);
   await toggleFinderFilesVisibility(!currentState);
   await LocalStorage.setItem(LocalStorageKey.LOCAL_HIDE_TOGGLE, !currentState);
 };