+++ extensions/hide-files/src/unhide-all-files.tsx
@@ -4,7 +4,7 @@ import { getFocusFinderPath } from "./utils/common-utils";
 import { removeFilesFromPanel } from "./utils/hide-files-utils";
 
 export default async () => {
-  await showHUD("Current window files are shown");
+  await showHUD("Unhidden Current window files");
   const finderPath = await getFocusFinderPath();
   const hideDesktopFilesCommand = `chflags nohidden ${finderPath.replace(" ", `" "`)}*`;
   exec(hideDesktopFilesCommand);