+++ extensions/hide-files/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## [Initial Version] - 2022-04-25
 
-- Hide and show files from any directory, including Desktop, Downloads, Documents, etc.
+- Hide and unhide files from any directory, including Desktop, Downloads, Documents, etc.